Change MaxFps type from int to float
Scrcpy 2.7 changed max_fps to accept float values (e.g. 29.97).
The server param is now passed as a float string instead of uint16.

+++ TqkLibrary.Scrcpy/Configs/VideoConfig.cs
@@ -30,7 +30,7 @@ namespace TqkLibrary.Scrcpy.Configs
         /// Default: 0 or null (unlimit)
         /// </summary>
         [OptionName("max_fps")]
-        public int MaxFps { get; set; } = 0;
+        public float MaxFps { get; set; } = 0;
         /// <summary>
         /// Max bitrate of video stream<br></br>
         /// Default: 0 or null (ignore)
